--- topic2/.babelrc
@@ -1,10 +1,0 @@
-{
-    "presets": [
-        [
-            "@babel/preset-env",
-            {
-              "useBuiltIns": "entry"
-            }
-          ]
-    ]
-}
