+++ topic2/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>Topic 2</title>
-    <script src="./script.js"></script>
+    <script src="/build/main.bundle.js"></script>
 </head>
 <body>
