+++ topic2/package.json
@@ -1,16 +1,31 @@
 {
-  "name": "topic2",
+  "name": "html5bootcamp",
   "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
+  "description": "##The future of Web Applications is here! ###Are you ready for it?",
+  "main": "modules/main.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "babel": "babel --presets es2015 modules/main.js -o dist/main.bundle.js",
+    "start": "http-server -p 8000",
+    "webpack": "webpack"
   },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/paulajofre/html5bootcamp.git"
+  },
+  "keywords": [],
   "author": "",
   "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/paulajofre/html5bootcamp/issues"
+  },
+  "homepage": "https://github.com/paulajofre/html5bootcamp#readme",
   "devDependencies": {
-    "@babel/cli": "^7.0.0",
-    "@babel/core": "^7.0.1",
-    "@babel/preset-env": "^7.0.0"
-  }
+    "babel-core": "^6.26.3",
+    "babel-loader": "^7.0.0",
+    "babel-preset-es2015": "^6.24.1",
+    "http-server": "^0.11.1",
+    "webpack": "^3.10.0",
+    "webpack-cli": "^3.1.0"
+  },
+  "dependencies": {}
 }
