+++ topic2/webpack.config.js
@@ -1,0 +1,24 @@
+var path = require('path');
+var webpack = require('webpack');
+
+ module.exports = {
+     entry: './modules/main.js',
+     output: {
+         path: path.resolve(__dirname, 'build'),
+         filename: 'main.bundle.js'
+     },
+     module: {
+        rules: [
+             {
+                 test: /\.js$/,
+                 loader: 'babel-loader',
+                 query: {
+                     presets: ['es2015']
+                 }
+             }
+         ]
+     },
+     stats: {
+         colors: true
+     }
+ };
